+++ releases/changes.md
@@ -1,3 +1,7 @@
-* Updated tests
-* Fixed bug in `fu-orf` to allow for single sequences
-* `fu-orf` also prints frame in the sequence comment
+
+* **fu-orf**
+  * Fixed bug in `fu-orf` to allow for single sequences
+  * Introduced `-r`, `--scan-reverse` to include reverse complement in the ORF finder
+  * `fu-orf` also prints frame in the sequence comment
+* Expanded test suite
+